+++ orsopy/__init__.py
@@ -1,3 +1,3 @@
 """Top-level package for orsopy."""
 
-__version__ = "1.2.2"
+__version__ = "1.2.3"